requirements.txt
clarifai==2.0.32
configparser==3.5.0
Django==1.11.6
django-taggit==0.22.1
enum34==1.1.6
future==0.15.2
jsonschema==2.5.1
kloudless==1.1.2
olefile==0.44
Pillow==2.9.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.13.0
six==1.11.0
